requirements.txt
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1